+++ tests/testthat/test-tbl_continuous.R
@@ -17,6 +17,19 @@ test_that("tbl_continuous() works", {
     c("**Characteristic**", "Drug A, N = 200", "Drug B, N = 200")
   )
 
+  expect_equal(
+    tbl_continuous(
+      data = trial[c("age", "grade")],
+      variable = age,
+      include = grade
+    ) %>%
+      `[`(c("table_body", "table_styling")),
+    trial %>%
+      select(age, grade) %>%
+      tbl_continuous(variable = age) %>%
+      `[`(c("table_body", "table_styling"))
+  )
+
   expect_error(
     tbl2 <-
       tbl_continuous(